Revise variable naming
While working on this branch it occurred to me that there are two times we're waiting for the manifest -- when we're fetching it directly from Bungie and when we're fetching it into state from local storage.  I wanted to start tracking the latter in state to make it a little clearer what our loading screen logic is assessing, which meant I needed to deal with 'manifest is loaded' not clearly expressing which kind of waiting for the manifest it's tracking.  The new language still feels a little clunky but is hopefully a little more expressive.