Show damage type of weapons
This one's a doozy.  Created an interface for the DestinyDamageTypeDefinition manifest table; added that table to the locally stored slice of the manifest; passed the item instance from the character component down to the item component; used the current damage type property in the item instance to do a lookup on the damage type definition table; and finally added the image of the damage type's icon back to the item component.