Change useManifestStatus output to string
useManifestStatus had been returning a boolean ('is the manifest loaded?') but there are really three states I wanted to capture: performing the version check, downloading the manifest if needed, and confirmation.  This allows the app to be a little more expressive with the user in describing what's happening during loading.