Handle bad response from Bungie API
This is a bandaid fix that hopefully handles the majority of instances where a user would experience an issue with the site during a maintenance window.  Every Bungie API request after initial page load retains dangerous assumptions re: always getting good data, but the entire fetching paradigm is currently up in the air for some high-level changes, so we're going to call a bandaid good enough for now.